+++ lib/rubocop/cop/rails/has_many_or_has_one_dependent.rb
@@ -23,6 +23,11 @@ module RuboCop
       class HasManyOrHasOneDependent < Cop
         MSG = 'Specify a `:dependent` option.'.freeze
 
+        def_node_matcher :activerecord_class, <<-PATTERN
+          {(const nil? :ApplicationRecord)
+           (const (const nil? :ActiveRecord) :Base)}
+        PATTERN
+
         def_node_matcher :association_without_options?, <<-PATTERN
           (send nil? {:has_many :has_one} _)
         PATTERN
@@ -46,7 +51,18 @@ module RuboCop
             (args) ...)
         PATTERN
 
-        def on_send(node)
+        def on_class(node)
+          _class_name, base_class, body = *node.children
+
+          activerecord_class(base_class) do
+            check_offsenses(body)
+            body.each_descendant(:send) do |n|
+              check_offsenses(n)
+            end
+          end
+        end
+
+        def check_offsenses(node)
           unless association_without_options?(node)
             return if valid_options?(association_with_options?(node))
           end