+++ spec/rubocop/cop/rails/has_many_or_has_one_dependent_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe RuboCop::Cop::Rails::HasManyOrHasOneDependent do
   context 'has_one' do
     it 'registers an offense when not specifying any options' do
       expect_offense(<<-RUBY.strip_indent)
-        class Person
+        class Person < ApplicationRecord
           has_one :foo
           ^^^^^^^ Specify a `:dependent` option.
         end
@@ -15,7 +15,7 @@ RSpec.describe RuboCop::Cop::Rails::HasManyOrHasOneDependent do
 
     it 'registers an offense when missing an explicit `:dependent` strategy' do
       expect_offense(<<-RUBY.strip_indent)
-        class Person
+        class Person < ApplicationRecord
           has_one :foo, class_name: 'bar'
           ^^^^^^^ Specify a `:dependent` option.
         end
@@ -24,7 +24,7 @@ RSpec.describe RuboCop::Cop::Rails::HasManyOrHasOneDependent do
 
     it 'does not register an offense when specifying `:dependent` strategy' do
       expect_no_offenses(<<-RUBY.strip_indent)
-        class Person
+        class Person < ApplicationRecord
           has_one :foo, dependent: :destroy
         end
       RUBY
@@ -33,7 +33,7 @@ RSpec.describe RuboCop::Cop::Rails::HasManyOrHasOneDependent do
     context 'with :through option' do
       it 'does not register an offense for non-nil value' do
         expect_no_offenses(<<-RUBY.strip_indent)
-          class Person
+          class Person < ApplicationRecord
             has_one :foo, through: :bar
           end
         RUBY
@@ -41,7 +41,7 @@ RSpec.describe RuboCop::Cop::Rails::HasManyOrHasOneDependent do
 
       it 'registers an offense for nil value' do
         expect_offense(<<-RUBY.strip_indent)
-        class Person
+        class Person < ApplicationRecord
           has_one :foo, through: nil
           ^^^^^^^ Specify a `:dependent` option.
         end
@@ -52,7 +52,7 @@ RSpec.describe RuboCop::Cop::Rails::HasManyOrHasOneDependent do
     context 'with_options dependent: :destroy' do
       it 'does not register an offense' do
         expect_no_offenses(<<-RUBY.strip_indent)
-          class Person
+          class Person < ApplicationRecord
             with_options dependent: :destroy do
               has_one :foo
             end
@@ -62,7 +62,7 @@ RSpec.describe RuboCop::Cop::Rails::HasManyOrHasOneDependent do
 
       it 'does not register an offense for using `class_name` option' do
         expect_no_offenses(<<-RUBY.strip_indent)
-          class Person
+          class Person < ApplicationRecord
             with_options dependent: :destroy do
               has_one :foo, class_name: 'Foo'
             end
@@ -75,7 +75,7 @@ RSpec.describe RuboCop::Cop::Rails::HasManyOrHasOneDependent do
   context 'has_many' do
     it 'registers an offense when not specifying any options' do
       expect_offense(<<-RUBY.strip_indent)
-        class Person
+        class Person < ApplicationRecord
           has_many :foo
           ^^^^^^^^ Specify a `:dependent` option.
         end
@@ -84,7 +84,7 @@ RSpec.describe RuboCop::Cop::Rails::HasManyOrHasOneDependent do
 
     it 'registers an offense when missing an explicit `:dependent` strategy' do
       expect_offense(<<-RUBY.strip_indent)
-        class Person
+        class Person < ApplicationRecord
           has_many :foo, class_name: 'bar'
           ^^^^^^^^ Specify a `:dependent` option.
         end
@@ -102,7 +102,7 @@ RSpec.describe RuboCop::Cop::Rails::HasManyOrHasOneDependent do
 
       it 'registers an offense for nil value' do
         expect_offense(<<-RUBY.strip_indent)
-        class Person
+        class Person < ApplicationRecord
           has_many :foo, through: nil
           ^^^^^^^^ Specify a `:dependent` option.
         end
@@ -113,7 +113,7 @@ RSpec.describe RuboCop::Cop::Rails::HasManyOrHasOneDependent do
     context 'Surrounded `with_options` block' do
       it 'registers an offense when `dependent: :destroy` is not present' do
         expect_offense(<<-RUBY.strip_indent)
-          class Person
+          class Person < ApplicationRecord
             with_options through: nil do
               has_many :foo
               ^^^^^^^^ Specify a `:dependent` option.
@@ -124,7 +124,7 @@ RSpec.describe RuboCop::Cop::Rails::HasManyOrHasOneDependent do
 
       it "doesn't register an offense for `with_options dependent: :destroy`" do
         expect_no_offenses(<<-RUBY.strip_indent)
-          class Person
+          class Person < ApplicationRecord
             with_options dependent: :destroy do
               has_many :foo
             end
@@ -136,7 +136,7 @@ RSpec.describe RuboCop::Cop::Rails::HasManyOrHasOneDependent do
         it "doesn't register an offense for " \
            '`with_options dependent: :destroy`' do
           expect_no_offenses(<<-RUBY.strip_indent)
-            class Person
+            class Person < ApplicationRecord
               with_options dependent: :destroy do
                 has_many :foo
                 has_many :bar
@@ -160,6 +160,33 @@ RSpec.describe RuboCop::Cop::Rails::HasManyOrHasOneDependent do
           end
         RUBY
       end
+    end
+  end
+
+  context 'base-class check' do
+    it 'runs the cop for ActiveRecord::Base class' do
+      expect_offense(<<-RUBY.strip_indent)
+        class Person < ActiveRecord::Base
+          has_one :foo
+          ^^^^^^^ Specify a `:dependent` option.
+        end
+      RUBY
+    end
+
+    it 'does not run the cop for ActiveResource class' do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        class Person < ActiveResource
+          has_one :foo
+        end
+      RUBY
+    end
+
+    it 'does not run the cop no base-class' do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        class Person
+          has_one :foo
+        end
+      RUBY
     end
   end
 end